backend/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.2
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.5
python-dotenv==1.0.1
python-multipart==0.0.6
bcrypt==4.2.1
requests==2.32.3
